refactor : Edit functions to read easily
